Move handlers to markdown directory
