+++ src/main/kotlin/com/framstag/taskdown/markdown/filehandler/MarkdownHistoryHandlerMap.kt
@@ -1,4 +1,4 @@
-package com.framstag.taskdown.database.filehandler
+package com.framstag.taskdown.markdown.filehandler
 
 val markdownHistoryHandlerMap = mapOf(
     LogHandler.NAME to LogHandler()
