+++ src/main/kotlin/com/framstag/taskdown/markdown/filehandler/MarkdownPropertyHandlerMap.kt
@@ -1,4 +1,4 @@
-package com.framstag.taskdown.database.filehandler
+package com.framstag.taskdown.markdown.filehandler
 
 val markdownPropertyHandlerMap = mapOf(
     TagsHandler.NAME to TagsHandler(),
