+++ src/test/kotlin/com/framstag/taskdown/markdown/filehandler/DueDateHandlerTest.kt
@@ -1,6 +1,6 @@
-package com.framstag.taskdown.database
+package com.framstag.taskdown.markdown.filehandler
 
-import com.framstag.taskdown.database.filehandler.DueDateHandler
+import com.framstag.taskdown.database.FileFormatException
 import com.framstag.taskdown.domain.TaskAttributes
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
