+++ src/test/kotlin/com/framstag/taskdown/markdown/filehandler/PriorityHandlerTest.kt
@@ -1,6 +1,6 @@
-package com.framstag.taskdown.database
+package com.framstag.taskdown.markdown.filehandler
 
-import com.framstag.taskdown.database.filehandler.PriorityHandler
+import com.framstag.taskdown.database.FileFormatException
 import com.framstag.taskdown.domain.Priority
 import com.framstag.taskdown.domain.TaskAttributes
 import org.junit.jupiter.api.Assertions
